+++ app/src/main/java/com/example/dsamo/foodmanager/models/database/entity/FridgeListBuying.java
@@ -12,24 +12,24 @@ import static android.arch.persistence.room.ForeignKey.CASCADE;
         foreignKeys = {@ForeignKey(entity = Fridge.class, parentColumns = "id", childColumns = "fridge_id", onDelete =  CASCADE),
         @ForeignKey(entity = ItemOfList.class, parentColumns = "id", childColumns = "list_product_id", onDelete =  CASCADE)})
 public class FridgeListBuying {
-    private int list_product_id;
-    private int fridge_id;
+    private long list_product_id;
+    private long fridge_id;
 
     public FridgeListBuying(){}
 
-    public int getList_product_id() {
+    public long getList_product_id() {
         return list_product_id;
     }
 
-    public void setList_product_id(int list_product_id) {
+    public void setList_product_id(long list_product_id) {
         this.list_product_id = list_product_id;
     }
 
-    public int getFridge_id() {
+    public long getFridge_id() {
         return fridge_id;
     }
 
-    public void setFridge_id(int fridge_id) {
+    public void setFridge_id(long fridge_id) {
         this.fridge_id = fridge_id;
     }
 }
